Fixed triple tests.
