+++ config/roboy.yml
@@ -29,7 +29,7 @@ config:
 ---
 module: genqa
 config:
-  genqa_server_address: "http://localhost:5000"
+  drqa_server_address: "http://localhost:5000"
   roboy_answer_sanity: 1000
 
 ---
