+++ modules/ravestate_nlp/__init__.py
@@ -2,106 +2,32 @@
 from ravestate import registry
 from ravestate.property import PropertyBase
 from ravestate.state import state
-from ravestate_nlp.question_words import QuestionWord
-from ravestate_nlp.rdf_triple import Triple
+from ravestate_nlp.question_word import QuestionWord
+from ravestate_nlp.triple import Triple
+from ravestate_nlp.extract_triples import extract_triples
 from ravestate.state import Emit
-
-
 import spacy
-from spacy.tokens import Token
+
 
 from reggol import get_logger
 logger = get_logger(__name__)
 
-SUBJECT_SET = {'nsubj'}
-OBJECT_SET = {'dobj', 'attr', 'advmod'}
-PREDICATE_SET = {'ROOT', 'conj'}
-PREDICATE_AUX_SET = {'acomp', 'aux', 'xcomp'}
-VERB_AUX_SET = {'VERB', 'ADJ'}
-
 
 def init_model():
-    global nlp
+    # TODO: Create nlp instance in :startup state, save in context instead of global var
+    global nlp, empty_token
     nlp = spacy.load('en_core_web_sm')
+    empty_token = nlp(u' ')[0]
+
+    # TODO: Make agent id configurable, rename nlp:contains-roboy to nlp:agent-mentioned
     about_roboy = ('you', 'roboy', 'robot', 'roboboy', 'your')
-    roboy_getter = lambda doc: any(roboy in doc.text.lower() for roboy in about_roboy)
+    def roboy_getter(doc) -> bool:
+        return any(roboy in doc.text.lower() for roboy in about_roboy)
+
     from spacy.tokens import Doc
     Doc.set_extension('about_roboy', getter=roboy_getter)
-    Doc.set_extension('triple', getter=extract_triples)
-
-
-def extract_triples(doc):
-    """
-    triple: subject, predicate, object
-    finding the triple of a sentence:
-    this is done by using the dependency tree of the sentence
-    a dependency tree shows the relation between the words of a sentence
-    the finite verb (predicate) is the structural center of the sentence and therefor of the tree
-    so starting with the predicate the algorithm searches through the dependency tree to find subject and object
-    -> triple_search()
-
-    examples:
-    - How are you?
-     tree depth = 1     predicate: "are"
-                            -> relation to "how": adverbial modifier (advmod) => object of this sentence
-                            -> relation to "you": nominal subject (nsubj) => subject of this sentence
-
-    - How old are you?
-    tree depth = 2      predicate: "are"
-                            -> relation to "old": adjectival complement (acomp) => auxiliary predicate
-                                    -> relation to "how":  adverbial modifier (advmod) => object of this sentence
-                            -> relation to "you": nominal subject (nsubj) => subject of this sentence
-
-    - Who is your father?
-    tree depth = 2      predicate: "is"
-                            -> relation to "who": nominal subject (nsubj) => subject of this sentence
-                            -> relation to "father": attribute (attr) => object of this sentence
-                                    ->relation to "your": possession modifier (poss)
-    """
-    triples = []
-    for token in doc:
-        if token.dep_ in PREDICATE_SET:
-            triple = triple_search(Triple(predicate=token), token)
-            triples.append(triple)
-    return triples
-
-
-def triple_search(triple: Triple, token: Token):
-    """
-    Recursive search through the dependency tree
-    looks for triple values in each of the children and calls itself with the children nodes
-    """
-    question_word = None
-    for word in token.children:
-        if word.text.lower() in QuestionWord.question_words:
-            question_word = QuestionWord(word)
-            word = QuestionWord(word)
-            if not triple.get_object() or triple.get_object().text == " ":
-                triple.set_object(question_word)
-        elif word.dep_ in OBJECT_SET:
-            triple.set_object(word)
-        if word.dep_ in SUBJECT_SET:
-            triple.set_subject(word)
-        if word.dep_ in PREDICATE_AUX_SET and word.pos_ in VERB_AUX_SET:
-            triple.set_predicate_aux(word)
-        if isinstance(word, Token):
-            triple = triple_search(triple, word)
-    if not triple.get_subject() and question_word: 
-        triple.set_subject(question_word)
-
-    # do not allow empty triples
-    empty_doc = nlp(u' ')
-    empty_token = empty_doc[0]
-    if not triple.get_subject():
-       triple.set_subject(empty_token)
-    if not triple.get_predicate():
-        triple.set_predicate(empty_token) 
-    if not triple.get_predicate_aux():
-        triple.set_predicate_aux(empty_token)
-    if not triple.get_object():
-        triple.set_object(empty_token)  
-        
-    return triple
+    Doc.set_extension('empty_token', getter=lambda doc: empty_token)
+    Doc.set_extension('triples', getter=extract_triples)
 
 
 @state(read="rawio:in", write=("nlp:tokens", "nlp:postags", "nlp:lemmas", "nlp:tags", "nlp:ner", "nlp:triples", "nlp:roboy","nlp:triples"))
@@ -128,7 +54,7 @@ def nlp_preprocess(ctx):
     ctx["nlp:ner"] = nlp_ner
     logger.info(f"[NLP:ner]: {nlp_ner}")
 
-    nlp_triples = nlp_doc._.triple
+    nlp_triples = nlp_doc._.triples
     ctx["nlp:triples"] = nlp_triples
     logger.info(f"[NLP:triples]: {nlp_triples}")
