+++ test/modules/ravestate_nlp/test_extract_triples.py
@@ -1,22 +1,21 @@
 import pytest
 import spacy
-from ravestate_nlp import extract_triples
-from ravestate_nlp.question_words import QuestionWord
+from ravestate_nlp import *
 
 
 @pytest.fixture
 def spacy_model():
     nlp = spacy.load('en_core_web_sm')
     from spacy.tokens import Doc
-    if not Doc.has_extension('triple'):
-        Doc.set_extension('triple', getter=extract_triples)
+    if not Doc.has_extension('triples'):
+        Doc.set_extension('triples', getter=extract_triples)
     return nlp
 
 
 def basic_test(expected_triples, spacy_model, text_input):
-    triples = spacy_model(text_input)._.triple
+    triples = spacy_model(text_input)._.triples
     print(triples)
-    assert triples == expected_triples
+    assert expected_triples == triples
 
 
 @pytest.mark.parametrize('text_input, expected_triples',
